Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,10 +2,10 @@
 #include<stdio.h>
 int main()
 {
-    int r;
+    float r;
     float area;
     printf("");
-    scanf("%d",&r);
+    scanf("%f",&r);
     area=3.14*r;
     printf("Area: %.2f",area);
 }
